Fix LD_LIBRARY_PATH and add -fastboot
LD_LIBRARY_PATH must be absolute and the variable must be exported.
This is now fixed using realpath(1) (part of the GNU core utilities).

Also, for certain games, `daphne -fastboot` will start much faster by
skipping the boot up sequence. That is now the default for those
games. It probably should be the default for all games, but I don't
know if it's harmful.